Some code cleanup to reorganize the project.
