+++ UnityProject/Assets/Scripts/DelegationUtils.cs
