+++ UnityProject/Assets/Scripts/PluginProxy.cs.meta
